Fixed Remove Product
Added entity reference field commerce_fulfilment_line_item to the package entity. Still need to change code to make-up for this. Package content should no longer work(easy to fix) and packing slips should no longer be correct (easy to fix.) MIGHT need to do the same for the shipping information. (Leaving it the way it is for now.)